refactor: move views into event-calendar/views/
