+++ apps/web/src/components/event-calendar/hooks/index.ts
@@ -11,7 +11,11 @@ export { useEventDialog } from "./use-event-dialog";
 export { useEventOperations } from "./use-event-operations";
 
 // View-specific hooks
-export { useEventCollection } from "./use-event-collection";
+export {
+  useEventCollection,
+  type EventCollectionForMonth,
+  type EventCollectionForWeek,
+} from "./use-event-collection";
 export { useGridLayout } from "./use-grid-layout";
 
 // Utility hooks
