+++ apps/web/src/components/event-calendar/index.ts
@@ -1,21 +1,19 @@
 "use client";
 
 // Component exports
-export { AgendaView } from "./agenda-view";
+export { AgendaView } from "./views/agenda-view";
 export { CalendarContent } from "./calendar-content";
 export { CalendarHeader } from "./calendar-header";
 export { CalendarNavigation } from "./calendar-navigation";
 export { CalendarViewSelector } from "./calendar-view-selector";
 export { CalendarViewTitle } from "./calendar-view-title";
-export { DayView } from "./day-view";
+export { DayView } from "./views/day-view";
 export { DraggableEvent } from "./draggable-event";
 export { DroppableCell } from "./droppable-cell";
 export { EventDialog } from "./event-dialog";
 export { EventItem } from "./event-item";
 export { EventsPopup } from "./events-popup";
 export { EventCalendar } from "./event-calendar";
-export { MonthView } from "./month-view";
-export { WeekView } from "./week-view";
 export { CalendarDndProvider, useCalendarDnd } from "./calendar-dnd-context";
 export { ViewPreferencesPopover } from "./view-preferences-popover";
