+++ apps/web/src/components/event-calendar/views/day-view.tsx
@@ -18,10 +18,10 @@ import {
   DroppableCell,
   EventItem,
   WeekCellsHeight,
-  useCurrentTimeIndicator,
   type CalendarEvent,
 } from "@/components/event-calendar";
 import { EndHour, StartHour } from "@/components/event-calendar/constants";
+import { useCurrentTimeIndicator } from "@/components/event-calendar/hooks";
 import { isMultiDayEvent } from "@/components/event-calendar/utils";
 import { cn } from "@/lib/utils";
